pyproject.toml
[tool.poetry]
name = "shadowbot"
version = "0.1.0"
description = ""
authors = ["eya46 <61458340+eya46@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
nonebot2 = { extras = ["fastapi", "httpx", "websockets"], version = "^2.2.1" }

nonebot-adapter-onebot = "^2.4.1"

nonebot-plugin-orm = { extras = ["asyncpg", "postgresql"], version = "^0.7.1" }
nonebot-plugin-apscheduler = "^0.4.0"
nonebot-plugin-alconna = "^0.38.2"
nonebot-plugin-chatrecorder = "^0.6.0"
nonebot-plugin-localstore = "^0.6.0"

arclet-alconna = "<1.8.4"

emoji = "^2.10.1"
jinja2 = "^3.1.3"
bilireq = "^0.2.12"
nb-cli = "^1.4.0"
wakeonlan = "^3.1.0"

#pydantic-extra-types = "^2.6.0"
pydantic = "~2"


[tool.nonebot]
plugins = ["nonebot_plugin_chatrecorder"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E402"]

line-length = 121